tests/wpt/tests/html/canvas/element/webgpu-access/2d.webgpuaccess.getTextureFormat.rgba8.tentative.https.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.webgpuaccess.getTextureFormat.rgba8.tentative.https</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.webgpuaccess.getTextureFormat.rgba8.tentative.https</h1>
<p class="desc">getTextureFormat() returns RGBA8 or BGRA8 for a typical context</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("getTextureFormat() returns RGBA8 or BGRA8 for a typical context");
_addTest(function(canvas, ctx) {

  assert_regexp_match(ctx.getTextureFormat(), /^rgba8unorm|bgra8unorm$/);

});
</script>
